+++ crates/pentect-cli/src/scan/engine.rs
@@ -2,9 +2,8 @@ use super::file_magic::{classify, FileMagic};
 use super::options::BinaryMode;
 use super::report::{FileFinding, ScanScope, SkippedFile};
 use super::walk::ignored_file_reason;
-use crate::infer_kind;
 use memchr::memchr;
-use pentect_core::{ByteRange, Category, Engine, Input, Kind, Profile, Span};
+use pentect_core::{infer_kind, ByteRange, Category, Engine, Input, Kind, Profile, Span};
 use std::collections::{BTreeMap, BTreeSet};
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicUsize, Ordering};
